+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### Unreleased
+* Add support for event notifications
+
 ### 5.5.0 / 2021-10-28
 * Add Component CRUD Support
 * Add Scheduler support