+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-### Unreleased
+### 5.6.0 / 2021-11-22
 * Add support for event notifications
 * Add more Scheduler support
 * Add metadata support for `Calendar`, `Message` and `Account`